django_mymedbook-master/django_mymedbook-master/backend/views/serials.py
from django.db import transaction

from rest_framework.views import APIView
from rest_framework.decorators import api_view, permission_classes, parser_classes
from rest_framework.parsers import FileUploadParser, MultiPartParser
from rest_framework.response import Response
from rest_framework.permissions import IsAuthenticated, AllowAny
from rest_framework.exceptions import ValidationError

from backend.serializers import *
from backend.models import *

class Serials(APIView):

    def get(self, request):
        pass