django_mymedbook-master/django_mymedbook-master/requirements.txt
Django==1.10.3
Pillow==4.0.0
PyJWT==1.4.2
argparse==1.2.1
django-braces==1.11.0
django-cors-headers==1.3.1
django-model-utils==2.6
django-oauth-toolkit==0.11.0
djangorestframework==3.5.3
djangorestframework-gis==0.11
djangorestframework-recursive==0.1.1
ipython==5.3.0
oauthlib==1.1.2
olefile==0.44
pathlib2==2.2.1
pickleshare==0.7.4
psycopg2==2.6.2
pyfcm==1.2.9
python-openid==2.2.5
python-social-auth==0.3.3
pytz==2016.10
requests==2.11.1
requests-oauthlib==0.8.0
requests-toolbelt==0.7.1
six==1.10.0
social-auth-core==1.2.0
traitlets==4.3.2
wsgiref==0.1.2
